Fix plugin export descriptor template type
